core: move retired root error to runtime
