+++ agent/plugins/errors.py
@@ -1,0 +1,5 @@
+"""Errors owned by the plugin runtime."""
+
+
+class RootRetired(RuntimeError):
+    """Report that a Root no longer accepts new work."""
