+++ tests/test_scheduler_v3_shadow.py
@@ -14,7 +14,7 @@ import pytest
 import agent.plugins.manager as plugin_manager_module
 from agent.control.timer import TimerReceipt, TimerStatus
 from agent.turn_effects import PostCommitEffect, TurnStorage
-from agent.control.scoped_turn import RootRetired
+from agent.plugins.errors import RootRetired
 from agent.plugin_composition.channels import ChannelDeliveryReceipt, DeliveryStatus
 from agent.plugin_composition.deliveries import PluginDeliveries
 from agent.plugin_composition.scoped_turns import PluginScopedTurns
